Update Big multiplication.cpp

+++ Big multiplication.cpp
@@ -4,24 +4,24 @@ using namespace std;
 
 int main()
 {
-    ll int  i,p;
+    ll i,p;
     cin >> p;
     for(i=0;i<p;i++)
     {
         string n,m;
-        ll int y = 0,u,g = 0;
+        ll y = 0,u,g = 0;
         cin >> n >> m;
-        ll int k = n.size();
-        long int l = m.size();
+        ll k = n.size();
+        ll l = m.size();
         for(int j=0;j<k;j++)
         {
             g = g + (n[j] - 48);
         }
-        for(int u=0;u<l;u++)
+        for(ll u=0;u<l;u++)
         {
             y = y + (m[u] - 48);
         }
-        ll int f = ((g%3)*(y%3))%3;
+        ll f = (((g%3)*(y%3))%3);
         cout << f << endl;
     }
     return 0;
